LayoutTests/platform/ios-simulator/fast/dom/custom/lifecycle-ready-parser-script-expected.txt

CONSOLE MESSAGE: line 9: TypeError: undefined is not a function (evaluating 'document.register("x-foo", { prototype: proto })')
This test ensures that the lifecycle callback of a parser-made element is visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL window.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
